@@ -2,9 +2,10 @@
   "name": "joeycf",
   "version": "0.3.14",
   "main": "./Core/index.js",
+  "type": "module",
   "description": "A framework for Cloudflare Workers",
   "scripts": {
-    "build": "npm run cleanup && tsc -p tsconfig-build.json && tsc -p bin/tsconfig.json && node scripts/create-dist-package-json.js",
+    "build": "npm run cleanup && tsc -p tsconfig.json && tsc -p bin/tsconfig.json && node scripts/create-dist-package-json.js",
     "build:prod": "npm run cleanup && tsc -p tsconfig-build.json && tsc -p bin/tsconfig.json",
     "prepub": "npm run build:prod && npm version patch && node scripts/create-dist-package-json.js",
     "prepub:minor": "npm run build:prod && npm version minor && node scripts/create-dist-package-json.js",
@@ -13,7 +14,7 @@
     "test": "c8 ava",
     "test:watch": "c8 ava --watch",
     "watch": "watch 'npm run build' ./src",
-    "cleanup": "rm -rf ./lib ./lib-test ./coverage",
+    "cleanup": "rm -rf ./lib ./coverage",
     "madge": "npm run build && mkdir -p .graphviz && madge --image .graphviz/src.svg --exclude '^(test)\\.ts$' src",
     "madge:lib": "npm run build && mkdir -p .graphviz && madge --extensions js --image .graphviz/lib.svg lib",
     "madge:lib-dts": "npm run build && mkdir -p .graphviz && madge --extensions ts --image .graphviz/lib-dts.svg lib"
@@ -57,13 +58,13 @@
     "@cloudflare/workers-types": "^3.14"
   },
   "devDependencies": {
-    "@ava/typescript": "^2.0.0",
+    "@ava/typescript": "^3.0.1",
     "@cloudflare/workers-types": "^3.14.1",
     "@types/lodash": "^4.14.178",
     "@types/node": "^17.0.24",
     "@typescript-eslint/eslint-plugin": "^5.21.0",
     "@typescript-eslint/parser": "^5.21.0",
-    "ava": "^3.15.0",
+    "ava": "^4.3.1",
     "c8": "^7.10.0",
     "camelcase": "^6.3.0",
     "cross-fetch": "^3.1.4",
@@ -73,10 +74,9 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-n": "^15.2.0",
     "eslint-plugin-promise": "^6.0.0",
-    "esm": "^3.2.25",
     "form-data": "^4.0.0",
     "madge": "^5.0.1",
-    "typescript": "^4.5.4",
+    "typescript": "^4.7.4",
     "watch": "^0.13.0",
     "wrangler": "^2.0.16"
   },
